Fix bugs https://github.com/foobnix/foobnix/issues/91 and https://github.com/foobnix/foobnix/issues/94
