+++ foobnix/util/id3_util.py
@@ -4,15 +4,18 @@ Created on 24 нояб. 2010
 
 @author: ivan
 '''
-from _bsddb import api
+
 import os
 import logging
 import urllib
 
+from mutagen.id3 import ID3
+from mutagen.flac import FLAC
+from mutagen.mp4 import MP4
+
 from gi.repository.GdkPixbuf import Pixbuf
 
 from zlib import crc32
-from subprocess import Popen, PIPE
 from tempfile import NamedTemporaryFile
 from foobnix.fc.fc import FC, FCache
 from foobnix.fc.fc_cache import COVERS_DIR
@@ -21,9 +24,7 @@ from foobnix.util.time_utils import convert_seconds_to_text
 from foobnix.util.bean_utils import update_bean_from_normalized_text
 from foobnix.util.file_utils import file_extension, get_file_extension
 from foobnix.util.audio import get_mutagen_audio
-from mutagen.id3 import ID3
-from mutagen.flac import FLAC
-from mutagen.mp4 import MP4
+
 
 RUS_ALPHABITE = "абвгдеёжзийклмнопрстуфхцчшщъыьэюя"
