+++ foobnix/version.py
@@ -1,1 +1,1 @@
-FOOBNIX_VERSION='3.1.00'
+FOOBNIX_VERSION='3.1.01'
