+++ po/ca.po
@@ -7,15 +7,15 @@ msgid ""
 msgstr ""
 "Project-Id-Version: foobnix\n"
 "Report-Msgid-Bugs-To: FULL NAME <EMAIL@ADDRESS>\n"
-"POT-Creation-Date: 2015-02-13 20:37+0300\n"
+"POT-Creation-Date: 2015-06-05 20:46+0300\n"
 "PO-Revision-Date: 2015-05-13 08:25+0000\n"
 "Last-Translator: Adolfo Jayme <fitoschido@gmail.com>\n"
 "Language-Team: Catalan <ca@li.org>\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: 8bit\n"
-"X-Launchpad-Export-Date: 2015-05-14 04:43+0000\n"
-"X-Generator: Launchpad (build 17493)\n"
+"X-Launchpad-Export-Date: 2015-06-06 05:15+0000\n"
+"X-Generator: Launchpad (build 17540)\n"
 
 #: .././foobnix/dm/dm.py:33 .././foobnix/preferences/preferences_window.py:58
 #: .././foobnix/gui/menu.py:55 .././foobnix/gui/top.py:44
@@ -634,11 +634,11 @@ msgstr ""
 msgid "Album title"
 msgstr "Títol de l'àlbum"
 
-#: .././foobnix/gui/coverlyrics.py:26
+#: .././foobnix/gui/coverlyrics.py:28
 msgid "Cover:"
 msgstr ""
 
-#: .././foobnix/gui/coverlyrics.py:33
+#: .././foobnix/gui/coverlyrics.py:34
 msgid "Lyric:"
 msgstr ""
 
@@ -674,27 +674,27 @@ msgstr "Informació"
 msgid "Downloads"
 msgstr "Baixades"
 
-#: .././foobnix/gui/infopanel.py:295
+#: .././foobnix/gui/infopanel.py:297
 msgid "Loading..."
 msgstr "S'està carregant..."
 
-#: .././foobnix/gui/infopanel.py:307
+#: .././foobnix/gui/infopanel.py:310
 msgid "The text not found"
 msgstr ""
 
-#: .././foobnix/gui/infopanel.py:344
+#: .././foobnix/gui/infopanel.py:347
 msgid "Similar Tags:"
 msgstr ""
 
-#: .././foobnix/gui/infopanel.py:355
+#: .././foobnix/gui/infopanel.py:358
 msgid "Similar Tracks:"
 msgstr ""
 
-#: .././foobnix/gui/infopanel.py:367
+#: .././foobnix/gui/infopanel.py:370
 msgid "Similar Artists:"
 msgstr ""
 
-#: .././foobnix/gui/infopanel.py:378
+#: .././foobnix/gui/infopanel.py:381
 msgid "Best Songs:"
 msgstr ""
